Modded get_latest_frame_moving_avg to have an int (not float) frame

+++ src/moving_avg_ticker.h
@@ -13,7 +13,7 @@ int reset_moving_avg(moving_avg_values * reset);
 
 int append_to_moving_avg(moving_avg_values *value, int appended, bool *isAtEnd);
 
-int get_latest_frame_moving_avg(moving_avg_values *value, float *frame);
+int get_latest_frame_moving_avg(moving_avg_values *value, int *frame);
 
 int free_moving_avg(moving_avg_values **value);
